Evolution: ColorBucket: Add a custom overlay preference controller [1/3]
- This handles loading of our custom overlays on dark theme.

Package naming format:
  - com.android.dark.themename (System)
  - com.android.dark.themenameExt (Settings)
  - com.android.dark.themenameExt (SystemUI)
  - com.android.dark.themenameExt (gboard....)

One of the overlays should be defined with the above package name format followed
by its theme name which will be considered as an initial or main overlay.
And the later overlays belonging to the same theme targeting different packages
should use the same package name as its main overlay but with "Ext" appended to the
end indicating it as an extension overlay and to be loaded along with the theme.

Ex: https://review.arrowos.net/c/ArrowOS/android_vendor_arrow/+/6459

This is just a simple overlay controller for now to put down the itch for theming
until we have a fully functioning aosp ThemePicker :p

Change-Id: I0bf38531d49ab8a66cf773807f3c114d22b38c45
Signed-off-by: ganeshi4u <ganeshvarmai4u@gmail.com>
Signed-off-by: Akito Mizukito <akito@evolution-x.org>
Signed-off-by: Joey Huab <joey@evolution-x.org>